requirements.txt
streamlit==1.51.0
chromadb==0.3.27
PyPDF2==3.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
google-generativeai
sentence-transformers==2.2.2